success

+++ philo_one/philo_one.h
@@ -22,7 +22,7 @@
 typedef struct s_philo
 {
 	int				num;
-	unsigned long	die;
+	int				die;
 	int				eat;
 	int				sleep;
 	int				times;
